FEAT : 오키친 관련 코드 초안 작성

+++ src/main/java/HeoJin/crawling_spring/entity/CookingOrder.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @Getter
 public class CookingOrder {
-    private Integer step;
+    private Long step;
     private String instruction;
 }
